requirements.txt
absl-py==0.11.0
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
Brlapi==0.6.6
cached-property==1.5.2
certifi==2018.1.18
chardet==3.0.4
click==6.7
cloudpickle==1.6.0
colorama==0.3.7
command-not-found==0.3
cryptography==2.1.4
cupshelpers==1.0
cycler==0.10.0
decorator==4.4.2
defer==1.0.6
distro-info===0.18ubuntu0.18.04.1
elementpath==2.0.5
future==0.18.2
gast==0.4.0
grpcio==1.34.0
gym==0.18.0
h5py==3.1.0
httplib2==0.9.2
idna==2.6
imageio==2.9.0
imgaug==0.4.0
importlib-metadata==3.3.0
joblib==1.0.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
loguru==0.5.3
louis==3.5.0
macaroonbakery==1.1.3
Mako==1.0.7
Markdown==3.3.3
MarkupSafe==1.0
matplotlib==3.3.3
mock==4.0.3
mpi4py==3.0.3
netifaces==0.10.4
networkx==2.5
numpy==1.19.4
oauth==1.0.1
olefile==0.45.1
opencv-python==4.5.1.48
pandas==1.2.0
pexpect==4.2.1
Pillow==7.2.0
pkg-resources==0.0.0
protobuf==3.14.0
py-trees==0.8.3
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pydot==1.4.1
pygame==2.0.1
pyglet==1.5.0
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
python-apt==1.6.5+ubuntu0.5
python-dateutil==2.8.1
python-debian==0.1.32
pytz==2018.3
PyWavelets==1.1.1
pyxdg==0.25
PyYAML==3.12
reportlab==3.4.0
requests==2.18.4
requests-unixsocket==0.1.5
scikit-image==0.18.1
scipy==1.6.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
Shapely==1.7.1
simplejson==3.13.2
six==1.15.0
system-service==0.3
systemd-python==234
tensorboard==1.13.1
tensorboardX==2.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
termcolor==1.1.0
tifffile==2020.12.8
torch==1.7.1
torchvision==0.8.2
tqdm==4.55.1
typing-extensions==3.7.4.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.22
usb-creator==0.3.3
wadllib==1.3.2
Werkzeug==1.0.1
xkit==0.0.0
xmlschema==1.4.1
zipp==3.4.0
zope.interface==4.3.2
